Bug fixes
This adds code to try the next metasource if we're unable to locate
a resource file. This is a normal occurence with langpacks for
multiple languages installed.

This also fixes a bug where we would keep an empty metasource around
after all of the sources were removed from it.

Finally, it adds metasource to the Debug implementation for FileSource.